add validation in view methods

+++ backend/categories/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render
-from django.http import HttpResponse
+from django.http import HttpResponse, HttpResponseBadRequest
 from models import Category
 
 def index(request):
@@ -10,15 +10,24 @@ def index(request):
 def create(request):
     """"""
     cname = request.POST['name']
+    if not cname:
+        return HttpResponseBadRequest('Category Name Required')
     category = Category(name=cname)
     category.save()
-    return HttpResponse('Category Created')
+    return HttpResponse('Category Created'), 201
 
 def edit(request, id):
     """"""
     cid = id
+    if not id:
+        return HttpResponseBadRequest("Invalid Category ID")
     category = Category.objects.get(id=cid)
-    category.delete()
+    if not category:
+        return HttpResponseBadRequest("Category not found")
+    cname = request.PUT['name']
+    if cname:
+        category.name = cname
+        category.save()
     return HttpResponse('Category edited')
 
 def find(request, id):
@@ -30,6 +39,10 @@ def find(request, id):
 def delete(request, id):
     """"""
     cid = id
+    if not id:
+        return HttpResponseBadRequest("Invalid Category ID")
     category = Category.objects.get(id=cid)
+    if not category:
+        return HttpResponseBadRequest("Category not found")
     category.delete()
     return HttpResponse('Category Deleted')
